fix wip_output.wav to wip_predict.py

+++ sound_file/wip_predict.py
@@ -29,14 +29,14 @@ while True:
     else:                                                  flag = False
 
     if (((time.time() - last_predict_time) > (1.0/rate)) and (not flag)):
-        wip_result = model.transcribe(os.path.join(get_package_share_directory('speech_recognition_whisper'), 'sound_file', 'wip_output.wav'))
-        # wip_result = {"text": str(random.randint(1,100))}
-        f = open(path, 'r', encoding='UTF-8')
-        data = f.read()
-        f.close()
-        f = open(path, 'w', encoding='UTF-8')
-        f.write("\n".join(data.split("\n")[:3]) + '\nTEXT:' + str(wip_result["text"]))
-        f.close()
+        if os.path.exists(os.path.join(get_package_share_directory('speech_recognition_whisper'), 'sound_file', 'wip_output.wav')):
+            wip_result = model.transcribe(os.path.join(get_package_share_directory('speech_recognition_whisper'), 'sound_file', 'wip_output.wav'))
+            f = open(path, 'r', encoding='UTF-8')
+            data = f.read()
+            f.close()
+            f = open(path, 'w', encoding='UTF-8')
+            f.write("\n".join(data.split("\n")[:3]) + '\nTEXT:' + str(wip_result["text"]))
+            f.close()
         last_predict_time = time.time()
     elif (not flag):
         time.sleep(0.1)
